File: crating methods to delete image from S3 on FileToS3 component

+++ src/main/java/com/rolemberg/eventostech/Services/Event/EventService.java
@@ -4,6 +4,7 @@ import com.rolemberg.eventostech.Domain.Event.Event;
 import com.rolemberg.eventostech.Domain.Event.EventCleanResponseDTO;
 import com.rolemberg.eventostech.Domain.Event.EventRegisterDTO;
 import com.rolemberg.eventostech.Domain.Event.EventsResponseDTO;
+import com.rolemberg.eventostech.Handlers.AppError;
 import com.rolemberg.eventostech.Repository.Event.EventRepo;
 import com.rolemberg.eventostech.Services.Address.AddressService;
 import com.rolemberg.eventostech.Util.FileToS3;
@@ -89,7 +90,9 @@ public class EventService {
     public Event deleteEvent(UUID id) throws NoSuchElementException {
         Event e = eventRepo
             .findById(id)
-            .orElseThrow();
+            .orElseThrow(() -> new AppError("Cannot get the event while trying to delete"));
+        boolean deleteFile = file.deleteFromS3(e.getImage_url());
+        if (!deleteFile) return null;
         eventRepo.deleteById(id);
         return e;
     }
